% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastmedian-package.R
\docType{package}
\name{fastmedian-package}
\alias{fastmedian}
\alias{fastmedian-package}
\title{fastmedian: Fast Median Function Using C++}
\description{
An optimzed median funciton using compiled C++ code.
}
\author{
\strong{Maintainer}: Matt Strimas-Mackey \email{mes335@cornell.edu} (\href{https://orcid.org/0000-0001-8929-7776}{ORCID})

Authors:
\itemize{
  \item Shawn Ligocki \email{sligocki@cornell.edu}
  \item Tom Auer \email{mta45@cornell.edu} (\href{https://orcid.org/0000-0001-8619-7147}{ORCID})
}

Other contributors:
\itemize{
  \item Cornell Lab of Ornithology [copyright holder]
}

}
\keyword{internal}
