% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{med}
\alias{med}
\title{Calculate the sample median using optimized C++ code}
\usage{
med(x)
}
\arguments{
\item{x}{numeric vector.}
}
\value{
Sample median of x after removing NA values.
}
\description{
Calculate the sample median using optimized C++ code
}
